2.18.3

+++ wp-health/src/Thirds/Cache/GlobalNginx.php
@@ -23,6 +23,16 @@ class GlobalNginx implements CacheCollectionItem
     public function clear()
     {
         try {
+            global $nginx_purger;
+
+            if (!isset($nginx_purger)) {
+                return false;
+            }
+
+            if (is_null($nginx_purger)) {
+                return;
+            }
+
             $nginx_purger->purge_all();
         } catch (\Exception $e) {
         }
